-changed mediaIds from uint32_t to uin16_t, making space for safe storage of the attribute filter param; this was using the last byte of d_fname[14] before, which meant it could be overwritten by software that clears d_fname (like e.g. BUGABOO)

+++ acsi2stm/Acsi.h
@@ -91,7 +91,7 @@ public:
   SdDev &blockDev;
 
   // Last media ID
-  uint32_t mediaId = 0;
+  uint16_t mediaId = 0;
 
   // SCSI status variables
   ScsiErr lastErr;
